+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+* Adds `mix_test` command with handy output
+* Adds `system_env` configuration for project setup
+
 ## 1.0.0
 
 Initial release